feat: added moviesection component, improved code structure
feat: added new env variable, variable for static server for images

feat: added new methods and updated their names

fix: removed duplciated code
