+++ env.d.ts
@@ -2,4 +2,5 @@ declare module '@env' {
   export const TMDB_ACCESS_TOKEN: string;
   export const TMDB_API_KEY: string;
   export const TMDB_URL: string;
+  export const TMDB_IMAGES_BASE_URL: string;
 }
