+++ src/components/CarouselHeader.tsx
@@ -25,15 +25,13 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     justifyContent: 'space-between',
     alignItems: 'center',
-    paddingHorizontal: 16,
-    marginBottom: 30,
     // borderWidth: 1,
     // borderColor: 'red',
   },
   title: {
     fontSize: 20,
     fontWeight: 'bold',
-    color: '#FFFFFF',
+    color: 'black',
     // borderWidth: 1,
     // borderColor: 'green',
   },
